app/requirements.txt
requests
requests[socks]
flask
openai
python-decouple
EdgeGPT
faster_whisper
GoogleBard
revChatGPT